app 이름 변경 및 클래스 속성 변경
