+++ students/junghoon/user/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class JunghoonConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'junghoon'
+    name = 'user'
