+++ students/junghoon/westagram/settings.py
@@ -41,7 +41,7 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'corsheaders',
     'django_extensions',
-    'junghoon',
+    'user',
 ]
 
 MIDDLEWARE = [
